+++ bsp/esp32_p4_function_ev_board/include/bsp/esp32_p4_function_ev_board.h
@@ -184,7 +184,7 @@ esp_err_t bsp_sdcard_unmount(void);
 
 #if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
 
-#define BSP_LCD_DRAW_BUFF_SIZE     (BSP_LCD_H_RES * 100) // Frame buffer size in pixels
+#define BSP_LCD_DRAW_BUFF_SIZE     (BSP_LCD_H_RES * 50) // Frame buffer size in pixels
 #define BSP_LCD_DRAW_BUFF_DOUBLE   (0)
 
 /**